backend/requirements.txt
Django>=5.2.4
psycopg2-binary
djangorestframework
